Move api.* to rest.*, correct error handling for not found
