+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/rest/NotFoundException.java
@@ -1,4 +1,4 @@
-package edu.kit.ipd.crowdcontrol.objectservice.api;
+package edu.kit.ipd.crowdcontrol.objectservice.rest;
 
 /**
  * Thrown on invalid requests.
