+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/rest/QueryParamUtil.java
@@ -1,4 +1,4 @@
-package edu.kit.ipd.crowdcontrol.objectservice.api;
+package edu.kit.ipd.crowdcontrol.objectservice.rest;
 
 import spark.Request;
